+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 
 - 完成实体的操作服务。
   - com.dwarfeng.familyhelper.life.impl.service.PbSetOperateServiceImpl。
+  - com.dwarfeng.familyhelper.life.impl.service.PbNodeOperateServiceImpl。
 
 #### Bug修复